requirements.txt
openai>=1.0.0
streamlit>=1.32.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0